+++ src/index.ts
@@ -13,7 +13,7 @@ export * as utils from "./lib/utils";
 export * as proto_message from "./proto/message";
 
 export * as waku from "./lib/waku";
-export { Waku, Protocols } from "./lib/waku";
+export { WakuNode, Protocols } from "./lib/waku";
 
 export * as waku_message from "./lib/waku_message";
 export { WakuMessage } from "./lib/waku_message";