+++ src/lib/enr/enr.node.spec.ts
@@ -2,8 +2,9 @@ import { expect } from "chai";
 
 import { makeLogFileName, NOISE_KEY_1, Nwaku } from "../../test_utils";
 import { createWaku } from "../create_waku";
+import type { Waku } from "../interfaces";
 import { waitForRemotePeer } from "../wait_for_remote_peer";
-import { Protocols, Waku } from "../waku";
+import { Protocols } from "../waku";
 
 import { ENR } from "./enr";