+++ src/lib/waku.ts
@@ -1,16 +1,19 @@
 import type { Stream } from "@libp2p/interface-connection";
 import type { PeerId } from "@libp2p/interface-peer-id";
+import type { PubSub } from "@libp2p/interface-pubsub";
 import { peerIdFromString } from "@libp2p/peer-id";
 import type { Multiaddr } from "@multiformats/multiaddr";
 import { multiaddr } from "@multiformats/multiaddr";
 import debug from "debug";
 import type { Libp2p } from "libp2p";
 
+import { Waku } from "./interfaces";
 import { FilterCodec, WakuFilter } from "./waku_filter";
 import { LightPushCodec, WakuLightPush } from "./waku_light_push";
 import { DecryptionMethod, WakuMessage } from "./waku_message";
 import { WakuRelay } from "./waku_relay";
 import { RelayCodecs, RelayPingContentTopic } from "./waku_relay/constants";
+import * as relayConstants from "./waku_relay/constants";
 import { StoreCodecs, WakuStore } from "./waku_store";
 
 export const DefaultPingKeepAliveValueSecs = 0;
@@ -43,12 +46,12 @@ export interface WakuOptions {
   decryptionKeys?: Array<Uint8Array | string>;
 }
 
-export class Waku {
+export class WakuNode implements Waku {
   public libp2p: Libp2p;
-  public relay: WakuRelay;
-  public store: WakuStore;
-  public filter: WakuFilter;
-  public lightPush: WakuLightPush;
+  public relay?: WakuRelay;
+  public store?: WakuStore;
+  public filter?: WakuFilter;
+  public lightPush?: WakuLightPush;
 
   private pingKeepAliveTimers: {
     [peer: string]: ReturnType<typeof setInterval>;
@@ -60,22 +63,27 @@ export class Waku {
   constructor(
     options: WakuOptions,
     libp2p: Libp2p,
-    store: WakuStore,
-    lightPush: WakuLightPush,
-    filter: WakuFilter
+    store?: WakuStore,
+    lightPush?: WakuLightPush,
+    filter?: WakuFilter
   ) {
     this.libp2p = libp2p;
-    this.relay = libp2p.pubsub as unknown as WakuRelay;
     this.store = store;
     this.filter = filter;
     this.lightPush = lightPush;
+
+    if (isWakuRelay(libp2p.pubsub)) {
+      this.relay = libp2p.pubsub;
+    }
+
     this.pingKeepAliveTimers = {};
     this.relayKeepAliveTimers = {};
 
     const pingKeepAlive =
       options.pingKeepAlive || DefaultPingKeepAliveValueSecs;
-    const relayKeepAlive =
-      options.relayKeepAlive || DefaultRelayKeepAliveValueSecs;
+    const relayKeepAlive = this.relay
+      ? options.relayKeepAlive || DefaultRelayKeepAliveValueSecs
+      : 0;
 
     libp2p.connectionManager.addEventListener("peer:connect", (evt) => {
       this.startKeepAlive(evt.detail.remotePeer, pingKeepAlive, relayKeepAlive);
@@ -179,9 +187,9 @@ export class Waku {
     key: Uint8Array | string,
     options?: { method?: DecryptionMethod; contentTopics?: string[] }
   ): void {
-    this.relay.addDecryptionKey(key, options);
-    this.store.addDecryptionKey(key, options);
-    this.filter.addDecryptionKey(key, options);
+    if (this.relay) this.relay.addDecryptionKey(key, options);
+    if (this.store) this.store.addDecryptionKey(key, options);
+    if (this.filter) this.filter.addDecryptionKey(key, options);
   }
 
   /**
@@ -191,9 +199,9 @@ export class Waku {
    * Strings must be in hex format.
    */
   deleteDecryptionKey(key: Uint8Array | string): void {
-    this.relay.deleteDecryptionKey(key);
-    this.store.deleteDecryptionKey(key);
-    this.filter.deleteDecryptionKey(key);
+    if (this.relay) this.relay.deleteDecryptionKey(key);
+    if (this.store) this.store.deleteDecryptionKey(key);
+    if (this.filter) this.filter.deleteDecryptionKey(key);
   }
 
   /**
@@ -229,11 +237,12 @@ export class Waku {
       }, pingPeriodSecs * 1000);
     }
 
-    if (relayPeriodSecs !== 0) {
+    const relay = this.relay;
+    if (relay && relayPeriodSecs !== 0) {
       this.relayKeepAliveTimers[peerIdStr] = setInterval(() => {
         log("Sending Waku Relay ping message");
         WakuMessage.fromBytes(new Uint8Array(), RelayPingContentTopic).then(
-          (wakuMsg) => this.relay.send(wakuMsg)
+          (wakuMsg) => relay.send(wakuMsg)
         );
       }, relayPeriodSecs * 1000);
     }
@@ -264,4 +273,17 @@ export class Waku {
     this.pingKeepAliveTimers = {};
     this.relayKeepAliveTimers = {};
   }
+}
+
+function isWakuRelay(pubsub: PubSub): pubsub is WakuRelay {
+  if (pubsub) {
+    try {
+      return pubsub.multicodecs.includes(
+        relayConstants.RelayCodecs[relayConstants.RelayCodecs.length - 1]
+      );
+      // Exception is expected if `libp2p` was not instantiated with pubsub
+      // eslint-disable-next-line no-empty
+    } catch (e) {}
+  }
+  return false;
 }