+++ src/lib/waku_filter/index.node.spec.ts
@@ -3,9 +3,10 @@ import debug from "debug";
 
 import { makeLogFileName, NOISE_KEY_1, Nwaku } from "../../test_utils";
 import { delay } from "../../test_utils/delay";
-import { createWaku } from "../create_waku";
+import { createFullNode } from "../create_waku";
+import type { WakuFull } from "../interfaces";
 import { waitForRemotePeer } from "../wait_for_remote_peer";
-import { Protocols, Waku } from "../waku";
+import { Protocols } from "../waku";
 import { WakuMessage } from "../waku_message";
 
 const log = debug("waku:test");
@@ -13,7 +14,7 @@ const log = debug("waku:test");
 const TestContentTopic = "/test/1/waku-filter";
 
 describe("Waku Filter", () => {
-  let waku: Waku;
+  let waku: WakuFull;
   let nwaku: Nwaku;
 
   afterEach(async function () {
@@ -25,7 +26,7 @@ describe("Waku Filter", () => {
     this.timeout(15000);
     nwaku = new Nwaku(makeLogFileName(this));
     await nwaku.start({ filter: true, lightpush: true });
-    waku = await createWaku({
+    waku = await createFullNode({
       staticNoiseKey: NOISE_KEY_1,
       libp2p: { addresses: { listen: ["/ip4/0.0.0.0/tcp/0/ws"] } },
     });
@@ -46,7 +47,7 @@ describe("Waku Filter", () => {
       expect(msg.payloadAsUtf8).to.eq(messageText);
     };
     await waku.filter.subscribe(callback, [TestContentTopic]);
-    // As the filter protocol does not cater for a ack of subscription
+    // As the filter protocol does not cater for an ack of subscription
     // we cannot know whether the subscription happened. Something we want to
     // correct in future versions of the protocol.
     await delay(200);