+++ src/lib/waku_store/index.node.spec.ts
@@ -7,14 +7,15 @@ import {
   NOISE_KEY_2,
   Nwaku,
 } from "../../test_utils";
-import { createWaku } from "../create_waku";
+import { createFullNode } from "../create_waku";
 import {
   generatePrivateKey,
   generateSymmetricKey,
   getPublicKey,
 } from "../crypto";
+import type { WakuFull } from "../interfaces";
 import { waitForRemotePeer } from "../wait_for_remote_peer";
-import { Protocols, Waku } from "../waku";
+import { Protocols } from "../waku";
 import { DecryptionMethod, WakuMessage } from "../waku_message";
 
 import { PageDirection } from "./history_rpc";
@@ -24,7 +25,7 @@ const log = debug("waku:test:store");
 const TestContentTopic = "/test/1/waku-store/utf8";
 
 describe("Waku Store", () => {
-  let waku: Waku;
+  let waku: WakuFull;
   let nwaku: Nwaku;
 
   afterEach(async function () {
@@ -48,7 +49,7 @@ describe("Waku Store", () => {
       ).to.be.true;
     }
 
-    waku = await createWaku({
+    waku = await createFullNode({
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();
@@ -81,7 +82,7 @@ describe("Waku Store", () => {
       ).to.be.true;
     }
 
-    waku = await createWaku({
+    waku = await createFullNode({
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();
@@ -121,7 +122,7 @@ describe("Waku Store", () => {
       ).to.be.true;
     }
 
-    waku = await createWaku({
+    waku = await createFullNode({
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();
@@ -158,7 +159,7 @@ describe("Waku Store", () => {
       ).to.be.true;
     }
 
-    waku = await createWaku({
+    waku = await createFullNode({
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();
@@ -201,7 +202,7 @@ describe("Waku Store", () => {
       ).to.be.true;
     }
 
-    waku = await createWaku({
+    waku = await createFullNode({
       pubSubTopic: customPubSubTopic,
       staticNoiseKey: NOISE_KEY_1,
     });
@@ -268,10 +269,10 @@ describe("Waku Store", () => {
     log("Messages have been encrypted");
 
     const [waku1, waku2, nimWakuMultiaddr] = await Promise.all([
-      createWaku({
+      createFullNode({
         staticNoiseKey: NOISE_KEY_1,
       }).then((waku) => waku.start().then(() => waku)),
-      createWaku({
+      createFullNode({
         staticNoiseKey: NOISE_KEY_2,
       }).then((waku) => waku.start().then(() => waku)),
       nwaku.getMultiaddrWithId(),
@@ -370,10 +371,10 @@ describe("Waku Store", () => {
     log("Messages have been encrypted");
 
     const [waku1, waku2, nimWakuMultiaddr] = await Promise.all([
-      createWaku({
+      createFullNode({
         staticNoiseKey: NOISE_KEY_1,
       }).then((waku) => waku.start().then(() => waku)),
-      createWaku({
+      createFullNode({
         staticNoiseKey: NOISE_KEY_2,
       }).then((waku) => waku.start().then(() => waku)),
       nwaku.getMultiaddrWithId(),
@@ -457,7 +458,7 @@ describe("Waku Store", () => {
       ).to.be.true;
     }
 
-    waku = await createWaku({
+    waku = await createFullNode({
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();